+++ lib/PhpParser/Node/Expr/Match_.php
@@ -3,17 +3,17 @@
 namespace PhpParser\Node\Expr;
 
 use PhpParser\Node;
-use PhpParser\Node\Stmt\Arm;
+use PhpParser\Node\MatchArm;
 
 class Match_ extends Node\Expr
 {
     /** @var Node\Expr */
     public $cond;
-    /** @var Arm[] */
+    /** @var MatchArm[] */
     public $arms;
 
     /**
-     * @param Arm[] $arms
+     * @param MatchArm[] $arms
      */
     public function __construct(Node\Expr $cond, array $arms = [], array $attributes = []) {
         $this->attributes = $attributes;
