+++ test/code/parser/expr/match.test
@@ -16,7 +16,7 @@ array(
                     value: 1
                 )
                 arms: array(
-                    0: Stmt_Arm(
+                    0: MatchArm(
                         condList: array(
                             0: Scalar_LNumber(
                                 value: 0
@@ -26,7 +26,7 @@ array(
                             value: Foo
                         )
                     )
-                    1: Stmt_Arm(
+                    1: MatchArm(
                         condList: array(
                             0: Scalar_LNumber(
                                 value: 1
@@ -61,7 +61,7 @@ array(
                     value: 1
                 )
                 arms: array(
-                    0: Stmt_Arm(
+                    0: MatchArm(
                         condList: array(
                             0: Scalar_LNumber(
                                 value: 0
@@ -104,7 +104,7 @@ array(
                     name: operator
                 )
                 arms: array(
-                    0: Stmt_Arm(
+                    0: MatchArm(
                         condList: array(
                             0: Expr_ClassConstFetch(
                                 class: Name(
@@ -151,7 +151,7 @@ array(
                     name: char
                 )
                 arms: array(
-                    0: Stmt_Arm(
+                    0: MatchArm(
                         condList: array(
                             0: Scalar_LNumber(
                                 value: 1
@@ -161,7 +161,7 @@ array(
                             value: 1
                         )
                     )
-                    1: Stmt_Arm(
+                    1: MatchArm(
                         condList: null
                         body: Scalar_String(
                             value: default
@@ -192,7 +192,7 @@ array(
                     value: 1
                 )
                 arms: array(
-                    0: Stmt_Arm(
+                    0: MatchArm(
                         condList: array(
                             0: Scalar_LNumber(
                                 value: 0
@@ -205,7 +205,7 @@ array(
                             value: Foo
                         )
                     )
-                    1: Stmt_Arm(
+                    1: MatchArm(
                         condList: null
                         body: Scalar_String(
                             value: Bar
